requirements.txt
cycler==0.11.0
fonttools==4.28.5
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.1
opencv-python==4.5.5.62
packaging==21.3
pandas==1.3.5
Pillow==9.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
